Address review: rename ElementDocEntry, simplify API, extract docs module
Rename BuiltinElementDocEntry to ElementDocEntry.
Fold the description into the Vec returned by element_doc_entries()
so the function returns a single Vec instead of (Option<String>, Vec).
Extract the docs submodule from load_builtins into its own file.

Strip ::: MDX admonition fences from hover tooltips
