+++ internal/compiler/langtype.rs
@@ -844,13 +844,13 @@ pub struct BuiltinElement {
     /// Documentation sections from builtins.slint, preserving source order.
     /// `Text` entries come from `///` (element-level) and `//!` (section) comments;
     /// `Member` entries reference a property, callback, or function by name.
-    pub docs: Vec<BuiltinElementDocEntry>,
+    pub docs: Vec<ElementDocEntry>,
 }
 
 /// One entry in the documentation of a builtin element, preserving the
 /// source order from `builtins.slint`.
 #[derive(Debug, Clone)]
-pub enum BuiltinElementDocEntry {
+pub enum ElementDocEntry {
     /// Free-form documentation text (from `///` or `//!` comments).
     Text(String),
     /// Reference to a property, callback, or function by name.
