+++ internal/compiler/load_builtins.rs
@@ -12,7 +12,7 @@ use std::rc::Rc;
 
 use crate::expression_tree::Expression;
 use crate::langtype::{
-    BuiltinElement, BuiltinElementDocEntry, BuiltinPrivateStruct, BuiltinPropertyDefault,
+    BuiltinElement, BuiltinPrivateStruct, BuiltinPropertyDefault,
     BuiltinPropertyInfo, DefaultSizeBinding, ElementType, Function, NativeClass, Type,
 };
 use crate::object_tree::{self, *};
@@ -189,7 +189,7 @@ pub(crate) fn load_builtins(register: &mut TypeRegister) {
         }
         properties
             .extend(builtin.native_class.properties.iter().map(|(k, v)| (k.clone(), v.clone())));
-        let (description, body) = docs::element_doc_entries(&c, &e);
+        let entries = docs::element_doc_entries(&c, &e);
         let parent_builtin = match &base {
             Base::NativeParent(p) => Some(p.as_ref()),
             _ => None,
@@ -197,7 +197,7 @@ pub(crate) fn load_builtins(register: &mut TypeRegister) {
         // Assemble docs as [description, inherited parent body, own body].
         // docs[0] is always the description so children can skip it
         // with `parent.docs[1..]`.
-        builtin.docs = docs::assemble(description, parent_builtin, body);
+        builtin.docs = docs::assemble(entries, parent_builtin);
 
         builtin.disallow_global_types_as_child_elements =
             parse_annotation("disallow_global_types_as_child_elements", &e).is_some();
@@ -302,160 +302,4 @@ fn parse_annotation(key: &str, node: &SyntaxNode) -> Option<Option<SmolStr>> {
     None
 }
 
-/// Extract `///` doc comments from the syntax tree of `builtins.slint`.
-mod docs {
-    use super::*;
-
-    /// Walk backwards across sibling tokens/nodes collecting consecutive
-    /// `///` doc comment lines immediately before `anchor`. Returns the
-    /// concatenated text with the `/// ` prefix stripped, or `None` if
-    /// no doc comment was present.
-    fn collect_before(anchor: &SyntaxNode) -> Option<String> {
-        let mut lines = Vec::new();
-        let mut cursor = anchor.node.prev_sibling_or_token();
-        while let Some(cur) = cursor {
-            match cur.kind() {
-                SyntaxKind::Whitespace => {}
-                SyntaxKind::Comment => {
-                    let text = cur.as_token().unwrap().text().to_string();
-                    if text.starts_with("///") {
-                        lines.push(text);
-                    } else if text.starts_with("//") {
-                        // Skip regular comments and //-annotations.
-                    } else {
-                        break;
-                    }
-                }
-                SyntaxKind::ExportsList => {
-                    // Doc comments may sit inside a preceding `export { ... }` list.
-                    if let Some(list) = cur.as_node() {
-                        let mut last = list.last_child_or_token();
-                        while let Some(child) = last {
-                            match child.kind() {
-                                SyntaxKind::Whitespace => {}
-                                SyntaxKind::Comment => {
-                                    let t = child.as_token().unwrap().text().to_string();
-                                    if t.starts_with("///") {
-                                        lines.push(t);
-                                    } else if t.starts_with("//") {
-                                        // skip
-                                    } else {
-                                        break;
-                                    }
-                                }
-                                _ => break,
-                            }
-                            last = child.prev_sibling_or_token();
-                        }
-                    }
-                    break;
-                }
-                _ => break,
-            }
-            cursor = cur.prev_sibling_or_token();
-        }
-        if lines.is_empty() {
-            return None;
-        }
-        lines.reverse();
-        Some(
-            lines
-                .iter()
-                .map(|t| t.strip_prefix("/// ").or_else(|| t.strip_prefix("///")).unwrap_or(""))
-                .collect::<Vec<_>>()
-                .join("\n"),
-        )
-    }
-
-    /// Extract the `///` doc comment before a syntax node. Also checks
-    /// above the enclosing `ExportsList` when the node is inside one.
-    pub(super) fn doc_comment(anchor: &SyntaxNode) -> Option<String> {
-        if let Some(doc) = collect_before(anchor) {
-            return Some(doc);
-        }
-        if let Some(parent) = anchor.parent()
-            && parent.kind() == SyntaxKind::ExportsList
-        {
-            return collect_before(&parent);
-        }
-        None
-    }
-
-    /// Extract the `///` description before the component and the ordered
-    /// body entries (`//!` text and member references) from inside it.
-    pub(super) fn element_doc_entries(
-        component: &SyntaxNode,
-        element: &syntax_nodes::Element,
-    ) -> (Option<String>, Vec<BuiltinElementDocEntry>) {
-        let description = doc_comment(component);
-
-        let mut entries = Vec::new();
-        let mut section_lines: Vec<String> = Vec::new();
-        let flush_section = |lines: &mut Vec<String>, entries: &mut Vec<BuiltinElementDocEntry>| {
-            if !lines.is_empty() {
-                entries.push(BuiltinElementDocEntry::Text(lines.join("\n")));
-                lines.clear();
-            }
-        };
-
-        for child in element.children_with_tokens() {
-            match child.kind() {
-                SyntaxKind::Comment => {
-                    if let Some(t) = child.as_token() {
-                        let text = t.text();
-                        if let Some(content) =
-                            text.strip_prefix("//! ").or_else(|| text.strip_prefix("//!"))
-                        {
-                            section_lines.push(content.to_string());
-                        }
-                    }
-                }
-                SyntaxKind::PropertyDeclaration => {
-                    let p = syntax_nodes::PropertyDeclaration::from(child.into_node().unwrap());
-                    if p.TwoWayBinding().is_some() {
-                        continue;
-                    }
-                    flush_section(&mut section_lines, &mut entries);
-                    let name = identifier_text(&p.DeclaredIdentifier()).unwrap();
-                    entries.push(BuiltinElementDocEntry::Member(name));
-                }
-                SyntaxKind::CallbackDeclaration => {
-                    let cb = syntax_nodes::CallbackDeclaration::from(child.into_node().unwrap());
-                    if cb.TwoWayBinding().is_some() {
-                        continue;
-                    }
-                    flush_section(&mut section_lines, &mut entries);
-                    let name = identifier_text(&cb.DeclaredIdentifier()).unwrap();
-                    entries.push(BuiltinElementDocEntry::Member(name));
-                }
-                SyntaxKind::Function => {
-                    let f = syntax_nodes::Function::from(child.into_node().unwrap());
-                    flush_section(&mut section_lines, &mut entries);
-                    let name = identifier_text(&f.DeclaredIdentifier()).unwrap();
-                    entries.push(BuiltinElementDocEntry::Member(name));
-                }
-                _ => {}
-            }
-        }
-        flush_section(&mut section_lines, &mut entries);
-        (description, entries)
-    }
-
-    /// Assemble the final doc entries for an element:
-    /// `[description, inherited parent body, own body]`.
-    pub(super) fn assemble(
-        description: Option<String>,
-        parent: Option<&BuiltinElement>,
-        body: Vec<BuiltinElementDocEntry>,
-    ) -> Vec<BuiltinElementDocEntry> {
-        let desc = description.unwrap_or_default();
-        let skip_inherited = desc.contains("\\skip_inherited");
-
-        let mut result = vec![BuiltinElementDocEntry::Text(desc)];
-        if !skip_inherited && let Some(parent) = parent {
-            result.extend(parent.docs[1..].iter().cloned());
-        }
-        result.extend(body);
-        result
-    }
-}
+use crate::doc_comments as docs;
